+++ src/consts.h
@@ -1,0 +1,10 @@
+#ifndef CONSTS_H
+#define CONSTS_H
+
+// Current Version: 1.1.0
+#define PHANTOMJS_VERSION_MAJOR  1
+#define PHANTOMJS_VERSION_MINOR  1
+#define PHANTOMJS_VERSION_PATCH  0
+#define PHANTOMJS_VERSION_STRING "PHANTOMJS_VERSION_MAJOR.PHANTOMJS_VERSION_MINOR.PHANTOMJS_VERSION_PATCH"
+
+#endif // CONSTS_H
