+++ src/phantomjs.pro
@@ -1,12 +1,23 @@
 TEMPLATE = app
 TARGET = phantomjs
-DESTDIR = ../bin
-HEADERS += csconverter.h
-SOURCES = phantomjs.cpp csconverter.cpp
-RESOURCES = phantomjs.qrc
 QT += network webkit
 CONFIG += console
-DEFINES += QT_NO_DEBUG_OUTPUT
+#DEFINES += QT_NO_DEBUG_OUTPUT
+
+DESTDIR = ../bin
+
+RESOURCES = phantomjs.qrc
+
+HEADERS += csconverter.h \
+    phantom.h \
+    webpage.h \
+    consts.h \
+    utils.h
+SOURCES += phantom.cpp \
+    webpage.cpp \
+    main.cpp \
+    csconverter.cpp \
+    utils.cpp
 
 include(gif/gif.pri)
